+++ src/@lib/equalities/index.ts
