+++ src/@lib/hooks/useMemo.ts
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
 import { DependencyList } from "react";
 import { shallowEquals } from "../equalities";
 import { useRef } from "./useRef";
@@ -17,12 +16,13 @@ export function useMemo<T>(
   }>({
     deps: [],
     value: undefined as unknown as T,
-    initialized: false
+    initialized: false,
   });
 
   // 2. 현재 의존성과 이전 의존성 비교
   // 초기화 상태이거나, 의존성이 변경된 경우 계산 실행
-  const depsChanged = !ref.current.initialized || !_equals(_deps, ref.current.deps);
+  const depsChanged =
+    !ref.current.initialized || !_equals(_deps, ref.current.deps);
 
   // 3. 의존성이 변경된 경우 factory 함수 실행 및 결과 저장
   if (depsChanged) {
@@ -33,5 +33,4 @@ export function useMemo<T>(
 
   // 4. 메모이제이션된 값 반환
   return ref.current.value;
-
 }
