+++ src/context/UserContext.tsx
@@ -1,0 +1,23 @@
+import { createContext, useContext } from "react";
+
+export interface User {
+  id: number;
+  name: string;
+  email: string;
+}
+
+export interface UserContextType {
+  user: User | null;
+  login: (email: string, password: string) => void;
+  logout: () => void;
+}
+
+export const UserContext = createContext<UserContextType | undefined>(undefined);
+
+export const useUserContext = () => {
+  const context = useContext(UserContext);
+  if (context === undefined) {
+    throw new Error("useAppContext must be used within an AppProvider");
+  }
+  return context;
+};
